\name{bookings}
\alias{bookings}
\docType{data}
\title{
  Hotel Bookings
}
\description{
  Numbers of hotel bookings made through a particular web-site
}
\usage{data(bookings)}
\format{
  A time series with 4344 observations.
}
\source{
  Kindly provided by Crimtan, UK for "ssadata" package.
}
\keyword{datasets}
