\name{dftreering}
\alias{dftreering}
\docType{data}
\title{Douglas Fir Tree Ring Indices}
\description{
  Douglas Fir, Snake River Basin, USA, 1282--1950
}
\usage{data(dftreering)}
\format{
  A time series with 669 observations.
}
\source{
  Hyndman, R.J. Time Series Data Library, http://data.is/TSDLdemo.

  Hydrology, Source: Hipel and McLeod (1994), in file: hurst/snake,
  Description: Douglas fir, snake River basin, U S A, 1282--1950
}
\keyword{datasets}
