\name{dwarfst}
\alias{dwarfst}
\docType{data}
\title{White Dwarf Star PG1159-035 Light Curve Data}
\description{
  A set of measurements of the light curve (time variation of the
  intensity) of the variable white dwarf star PG1159-035 during March 1989.
}
\usage{data(dwarfst)}

\format{A time series of length 618.}

\details{
  This is a set of measurements of the light curve (time variation of
  the intensity) of the variable white dwarf star PG1159-035 during
  March 1989. It was recorded by the Whole Earth Telescope (a
  coordinated group of telescopes distributed around the earth that
  permits the continuous observation of an astronomical object) and
  submitted by James Dixson and Don Winget of the Department of
  Astronomy and the McDonald Observatory of the University of Texas at
  Austin. The telescope is described in an article in The Astrophysical
  Journal (361), p. 309-317 (1990), and the measurements on PG1159-035
  will be described in an article scheduled for the September 1 issue of
  the Astrophysical Journal. The observations were made of PG1159-035
  and a nonvariable comparison star. A polynomial was fit to the light
  curve of the comparison star, and then this polynomial was used to
  normalize the PG1159-035 signal to remove changes due to varying
  extinction (light absorption) and differing telescope properties.
}

\note{
  Only the part I of the original dataset is included.
}

\source{
  The Santa Fe Time Series Competition Data
  \url{http://www-psych.stanford.edu/~andreas/Time-Series/SantaFe.html}
}

\references{
  A. S. Weigend and N. A. Gershenfeld, eds.. Time Series Prediction:
  Forecasting the Future and Understanding the Past. Reading, MA:
  Addison-Wesley, 1994.
}

\keyword{datasets}
