\name{g15}
\alias{g15}
\docType{data}
\title{
  GLONASS final time correction data
}
\description{
  GLONASS final time correction data for satellite 15 taken from 02/01/2014 to
  31/12/2014 within 5 minute intervals.
}
\usage{data(g15)}
\format{
  A time series with 104832 observations.
}
\note{
  There are total 895 missed points.
}
\source{
  \url{https://www.glonass-iac.ru/en/index.php}
}
\keyword{datasets}
