\name{hotel}
\alias{hotel}
\docType{data}
\title{Monthly Hotel Occupied Room Averages}
\description{
  Monthly hotel occupied room averages. From January 1963 till December 1976.
}
\usage{data(hotel)}
\format{
  A time series with 168 observations
}
\source{
  Hyndman, R.J. Time Series Data Library, http://data.is/TSDLdemo.
}
\keyword{datasets}
