\name{kruppel}
\alias{kruppel}
\docType{data}
\title{
  ``Krueppel'' Gene Expression Data
}
\description{
  Regularized data of gene expression for the ``Krueppel'' gene (a drosophila embryo)
}
\usage{data(kruppel)}
\format{
  A double matrix with integer values.
}
\details{
  The data was downloaded from the BDTNP archive (file
  v5_s11643-28no06-04.pca), and then was preprocessed to obtain a
  digital image by means of BioSSA (Shlemov et al, 2014).
}
\source{
  Berkeley Drosophila Transcription Network Project (2014)
  \url{http://bdtnp.lbl.gov/Fly-Net/}
}
\seealso{
  \url{http://biossa.github.io/example_application.html}
}
\references{
  Shlemov A, Golyandina N, Holloway D, Spirov A (2015) Shaped singular spectrum analysis
  for quantifying gene expression, with application to the early Drosophila embryo. BioMed
  Research International 2015 (Article ID 689745).
  
  Shlemov A, Golyandina N (2014) Shaped extensions of singular spectrum
  analysis \url{http://arxiv.org/pdf/1401.4980.pdf}
}
\keyword{datasets}
