\name{monet}
\alias{monet}
\docType{data}
\title{
  The painting of Claude Monet.
}
\description{
 A preview of the painting of Claude Monet called ``Study of a Figure
 Outdoors: Woman with a Parasol, facing left'', 1886.
}
\usage{data("monet")}
\details{
  Obtained from converting an 400 x 263 x 3 (color, RGB model) image.
}
\format{
  An array of rank 3.
}
\source{
  \url{https://commons.wikimedia.org/wiki/File:Monet.012.sonnenschirm.jpg}
}
\keyword{datasets}


