\name{oilproduction}
\alias{oilproduction}
\docType{data}
\title{Cruide Oil and Natural Gas Plant Liquids Production}
\description{
  Monthly cruide oil and natural gas plant liquids production from January 1973
  till December 1997
}
\usage{data(oilproduction)}
\format{
  A time series with 300 observations.
}
\note{
  Part of the table 1.2 ``Primary energy production by source'' from
  Monthly Energy Review (MER)
}
\source{
  U.S. Energy Information Administration \url{http://www.eia.gov/totalenergy/data/monthly/#summary}
}
\keyword{datasets}
