\name{paynsa}
\alias{paynsa}
\docType{data}
\title{
  All Employees: Total Nonfarm Payrolls
}
\description{
 Employees on nonfarm payrolls by industry sector and selected industry
 detail, Not seasonally adjusted. From January 1936 till January 2015.
}
\usage{data(paynsa)}
\format{
  A time series with 913 observations.
}
\source{
  US. Bureau of Labor Statistics, All Employees: Total Nonfarm Payrolls
  [PAYNSA], retrieved from TSDL (Hyndman, R.J. Time Series Data Library,
  http://data.is/TSDLdemo)
}
\keyword{datasets}
