\name{sunspot2}
\alias{sunspot2}
\docType{data}
\title{Yearly Mean Total Sunspot Number}
\description{  
  Yearly mean total sunspot number obtained by taking a simple arithmetic
  mean of the daily total sunspot number over all days of each year.
}
\usage{data(sunspot2)}
\format{
  A time series of 316 observations.
}
\note{
  In early years in particular before 1749, the means are computed on
  only a fraction of the days in each year because on many days, no
  observation is available.

  This is a 'V2' version of this dataset.
}
\source{
  WDC-SILSO, Royal Observatory of Belgium, Brussels.
  \url{http://www.sidc.be/silso/datafiles}

  Obtained from TSDL (Hyndman, R.J. Time Series Data Library,
  http://data.is/TSDLdemo)
}
\keyword{datasets}
